+++ packages/react-voting-app/package.json
@@ -5,10 +5,12 @@
   "dependencies": {
     "@bvotal/common": "1.0.0",
     "@material-ui/core": "^4.11.0",
+    "@types/react-router-dom": "^5.1.6",
     "fontsource-roboto": "^3.0.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-intl": "^5.8.5",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
     "typescript": "^4.0.3"
   },
