+++ packages/react-voting-app/src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
+import { BrowserRouter } from "react-router-dom";
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import 'fontsource-roboto';
@@ -9,7 +10,9 @@ import { IntlProvider } from 'react-intl';
 ReactDOM.render(
   <React.StrictMode>
     <IntlProvider locale='en'>
+    <BrowserRouter>
       <App />
+    </BrowserRouter>
     </IntlProvider>
   </React.StrictMode>,
   document.getElementById('root')
