+++ packages/react-voting-app/src/pages/start/StartPage.css
@@ -1,0 +1,39 @@
+.StartPage {
+    text-align: center;
+  }
+  
+  .StartPage-logo {
+    height: 40vmin;
+    pointer-events: none;
+  }
+  
+  @media (prefers-reduced-motion: no-preference) {
+    .StartPage-logo {
+      animation: StartPage-logo-spin infinite 20s linear;
+    }
+  }
+  
+  .StartPage-header {
+    background-color: #282c34;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+  }
+  
+  .StartPage-link {
+    color: #61dafb;
+  }
+  
+  @keyframes StartPage-logo-spin {
+    from {
+      transform: rotate(0deg);
+    }
+    to {
+      transform: rotate(360deg);
+    }
+  }
+
